Program not cuntional atm, just needed to update repo

+++ src/main.rs
@@ -1,20 +1,18 @@
-use std::io;
-use std::io::Write;
 use libtor::{Tor, TorFlag, TorAddress, HiddenServiceVersion};
-use std::process::Command;
-use std::thread;
-use std::time::Duration;
-
+use std::{io, io::Write, process::Command, thread, time::Duration};
+use std::net::{TcpListener, TcpStream};
 enum InputKind {
     Server{hs_port: u16, tor_port: u16},
     Username{username: String},
-    Misc{result1: u16}
+    Misc{result1: u16},
+    Message{message: String}
 }
 
 fn main(){
     
     println!("Welcome to TMMP(\"Tor Minimal Messaging Protocol\"), run /help");
-    let mut global_username = String::new();
+    let mut _global_username = String::new();
+    let mut _msg = String::new();
     loop{
 
         let x =  get_input();
@@ -39,43 +37,74 @@ fn main(){
                     println!("Run /join to join a server\n");
                     println!("Run /leave to leave a server\n");
                     println!("Run \"/server (port_for_hidden_service)\" To run a server(you will be unable to join a server until you stop yours). Max port number is 65,535 use no commas.");
-                    println!("Run /msg to send a message\n");
+                    println!("\nRun /msg to send a message\n");
                 }
             },
 
             InputKind::Username{username} => {
-                global_username = username;
-                println!("{}",global_username);
+                _global_username = username;
+                println!("{}",_global_username);
             },
 
             InputKind::Server { hs_port, tor_port } => {
-                
+                let port_limit: u16 = 65535;
                 println!("Creating server...");
-            
+                if hs_port == tor_port {
+                    println!("[system]$ ERROR: hidden servie and tor port are the same! Exitting...");
+                    break;
+                }
+
+                if hs_port > port_limit {
+                    println!("[system]$ ERROR: hidden service port is greater than 65,535! Exitting...");
+                    break;
+                }
+
+                if tor_port > port_limit {
+                    println!("[system]$ ERROR: tor service port is greater than 65,535! Exitting...");
+                    break;
+                }
                 command("rm -rf /tmp/tor-rust");
-                let thread2 = thread::spawn(|| {
-                    thread::sleep(Duration::from_secs(5));
-                    println!();
-                    println!("Starting server...");
                 
+                let thread2 = thread::spawn(move || {
+                    
+                    Tor::new()
+                            .flag(TorFlag::DataDirectory("/tmp/.tmmp-tor".into()))
+                            .flag(TorFlag::SocksPort(tor_port))
+                            .flag(TorFlag::HiddenServiceDir("/tmp/.tmmp-tor/hs-dir".into()))
+                            .flag(TorFlag::HiddenServiceVersion(HiddenServiceVersion::V3))
+                            .flag(TorFlag::HiddenServicePort(TorAddress::Port(hs_port),None.into()))
+                            .start_background();
                 });
-
-                Tor::new()
-                    .flag(TorFlag::DataDirectory("/tmp/.tmmp-tor".into()))
-                    .flag(TorFlag::SocksPort(tor_port))
-                    .flag(TorFlag::HiddenServiceDir("/tmp/.tmmp-tor/hs-dir".into()))
-                    .flag(TorFlag::HiddenServiceVersion(HiddenServiceVersion::V3))
-                    .flag(TorFlag::HiddenServicePort(TorAddress::Port(hs_port),None.into()))
-                    .start_background();
-
-            
+                
+                
+                if thread2.is_finished() == true {
+                    
+                        let thread1 = thread::spawn(move || {
+                            println!("[system]$Binding socket to port {}...",hs_port);
+                            let listener_port = "127.0.0.1:".to_string()+&hs_port.to_string();
+                            let listener = TcpListener::bind(listener_port);
+                        });
+                    
+                }
+                    
+                thread2.join().unwrap();
+                
+                
+                
+                
                 println!("Test");
 
-                thread2.join().unwrap();
+                
             
                 println!("\nstopping server\n");
 
-            }
+            },
+
+            InputKind::Message { message } => {
+
+                _msg = message.clone();
+
+            },
         };    
     }
 
@@ -96,7 +125,8 @@ fn get_input() -> InputKind {
         .expect("Error getting inputt");
 
     println!();
-
+    let len = input1.len();
+    let slice1 = input1[4..len].to_string();
     let v: Vec<&str> = input1.split(' ').collect();
 
     
@@ -137,9 +167,13 @@ fn get_input() -> InputKind {
             Err(..) => println!(),
         };
         
-
         return InputKind::Misc{result1: 0};
 
+    }
+
+    if v[0].trim() == "/msg" {
+
+        return InputKind::Message{message: slice1};
     }
 
     return InputKind::Misc{result1: 0};
